MODCAT-2 - Logical separation from DAO for model management
