+++ src/main/java/org/folio/cataloging/business/cataloguing/authority/AuthorityItem.java
@@ -15,8 +15,9 @@ import org.apache.commons.logging.LogFactory;
 
 import org.folio.cataloging.business.cataloguing.common.CatalogItem;
 import org.folio.cataloging.business.cataloguing.common.ItemEntity;
-import org.folio.cataloging.business.cataloguing.common.Model;
-import org.folio.cataloging.business.cataloguing.common.ModelItem;
+import org.folio.cataloging.dao.persistence.AuthorityModelItem;
+import org.folio.cataloging.dao.persistence.Model;
+import org.folio.cataloging.dao.persistence.ModelItem;
 import org.folio.cataloging.business.cataloguing.common.Tag;
 import org.folio.cataloging.business.cataloguing.common.TagImpl;
 import org.folio.cataloging.exception.MandatoryTagException;
