+++ src/main/java/org/folio/cataloging/business/cataloguing/bibliographic/BibliographicItem.java
@@ -12,7 +12,7 @@ import javax.xml.parsers.ParserConfigurationException;
 import org.folio.cataloging.bean.cataloguing.bibliographic.BibliographicEditBean;
 import org.folio.cataloging.business.cataloguing.common.CatalogItem;
 import org.folio.cataloging.business.cataloguing.common.ItemEntity;
-import org.folio.cataloging.business.cataloguing.common.ModelItem;
+import org.folio.cataloging.dao.persistence.ModelItem;
 import org.folio.cataloging.business.cataloguing.common.Tag;
 import org.folio.cataloging.business.cataloguing.common.TagImpl;
 import org.folio.cataloging.business.common.DataAccessException;
