+++ src/main/java/org/folio/cataloging/dao/BibliographicModelItemDAO.java
@@ -7,17 +7,21 @@ import net.sf.hibernate.Hibernate;
 import net.sf.hibernate.HibernateException;
 import net.sf.hibernate.Session;
 import net.sf.hibernate.type.Type;
-import org.folio.cataloging.business.cataloguing.bibliographic.BibliographicModelItem;
-
+import org.folio.cataloging.dao.persistence.BibliographicModelItem;
 
 /**
- * @author Wim Crols
+ * The Class BibliographicModelItemDAO.
+ *
+ * @author paulm
+ * @author carment
  * @since 1.0
  */
 public class BibliographicModelItemDAO extends ModelItemDAO {
 
-	/* (non-Javadoc)
-	 * @see DAOModelItem#getPersistentClass()
+	/**
+	 * Gets the persistent class.
+	 *
+	 * @return the persistent class
 	 */
 	protected Class getPersistentClass() {
 		return BibliographicModelItem.class;
@@ -27,13 +31,14 @@ public class BibliographicModelItemDAO extends ModelItemDAO {
 	 * Return true if the given model is used by an item
 	 *
 	 * @param bibItem the id of the model item
+	 * @param session the session
 	 * @return true if the given model is used by an item
 	 * @throws HibernateException in case of data access failure
 	 */
 public boolean getModelUsageByItem(int bibItem, final Session session)
 	throws HibernateException {
 	List <Integer> list =
-		find(
+			session.find(
 			"select count(*) from "
 				+ getPersistentClass().getName()
 				+ " as b"
