+++ src/main/resources/org/folio/cataloging/dao/persistence/AUTHORITY_MODEL_ITEM.hbm.xml
@@ -4,14 +4,14 @@
 	"http://hibernate.sourceforge.net/hibernate-mapping-2.0.dtd">
 <hibernate-mapping>
 	<class
-		name="org.folio.cataloging.business.cataloguing.authority.AuthorityModelItem"
+		name="org.folio.cataloging.dao.persistence.AuthorityModelItem"
 		table="AUTHORITY_MODEL_ITEM">
 		<id name="item">
 			<column name="ITEM" sql-type="NUMBER (11)" not-null="true"/>
 			<generator class="assigned"/>
 		</id>
 		<many-to-one name="model" column="MODEL" 
-		   class="org.folio.cataloging.business.cataloguing.authority.AuthorityModel"
+		   class="org.folio.cataloging.dao.persistence.AuthorityModel"
 		   not-null="true"/>
 		<property name="xmlFields" column="XML_FIELDS" 
 		   type="librisuite.business.common.hibernate.StringClobType"/>
